move most of the MPHF stuff into one file
