+++ spacegraphcats/search/__init__.py
@@ -1,2 +1,2 @@
-from .index import MPHF_KmerIndex
+from spacegraphcats.cdbg.index_contigs_by_kmer import MPHF_KmerIndex
 from .catlas import CAtlas
