+++ spacegraphcats/search/search_utils.py
@@ -9,7 +9,7 @@ from screed.utils import to_str
 from sourmash import MinHash
 
 from spacegraphcats.utils.bgzf.bgzf import BgzfReader
-from .index import MPHF_KmerIndex
+from . import MPHF_KmerIndex
 
 
 def sqlite_get_max_offset(cursor):
